App works, Validation and styles are still needed

+++ Application/Interfaces/IDepositCalculatorService.cs
@@ -4,5 +4,5 @@ namespace Application.Interfaces;
 
 public interface IDepositCalculatorService
 {
-    long CalculateDepositIncome(DepositDto depositDto);
+    double CalculateDepositIncome(DepositDto depositDto);
 }
